Title: 실수 받아 그대로 출력, Time: 131ms, Memory: 10MB, Status: Runtime Error - Codetree

+++ 260128/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -5,7 +5,7 @@ public class Main {
         // Please write your code here.
         Scanner sc=new Scanner(System.in);
 
-        float N=sc.nextInt();
+        double N=sc.nextInt();
 
         System.out.printf("%.2f",N);
     }
